requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
jinja2==3.1.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
aiofiles==24.1.0
openpyxl==3.1.5
apscheduler==3.10.4
httpx==0.27.0
itsdangerous==2.1.2
